requirements.txt
aiogram==3.11.0
aiohttp==3.9.5
python-dotenv==1.0.1